+++ packages/blocks/src/_specs/_specs.ts
@@ -41,7 +41,6 @@ import { AFFINE_SLASH_MENU_WIDGET } from '../page-block/widgets/slash-menu/index
 import { ParagraphBlockSchema } from '../paragraph-block/paragraph-model.js';
 import { ParagraphService } from '../paragraph-block/paragraph-service.js';
 import { SurfaceBlockSchema } from '../surface-block/surface-model.js';
-import { SurfacePageService } from '../surface-block/surface-page-service.js';
 import { SurfaceService } from '../surface-block/surface-service.js';
 import {
   SurfaceRefBlockSchema,
@@ -198,7 +197,7 @@ export const DocEditorBlockSpecs: BlockSpec[] = [
     view: {
       component: literal`affine-surface`,
     },
-    service: SurfacePageService,
+    service: SurfaceService,
   },
   {
     schema: SurfaceRefBlockSchema,
