+++ packages/blocks/src/surface-ref-block/surface-ref-renderer.ts
@@ -7,7 +7,7 @@ import type { EdgelessElement, TopLevelBlockModel } from '../_common/types.js';
 import type { NoteBlockModel } from '../note-block/index.js';
 import { Renderer } from '../surface-block/index.js';
 import type { SurfaceBlockModel } from '../surface-block/surface-model.js';
-import type { SurfacePageService } from '../surface-block/surface-page-service.js';
+import type { SurfaceService } from '../surface-block/surface-service.js';
 import { getSurfaceBlock } from './utils.js';
 
 type RefElement = Exclude<EdgelessElement, NoteBlockModel>;
@@ -27,7 +27,7 @@ export class SurfaceRefRenderer {
   };
 
   get surfaceService() {
-    return this.std.spec.getService('affine:surface') as SurfacePageService;
+    return this.std.spec.getService('affine:surface') as SurfaceService;
   }
 
   get surfaceRenderer() {
